Move shell fixture to helpers
