+++ spacecmd/tests/test_activationkey.py
@@ -8,20 +8,7 @@ import time
 import hashlib
 import spacecmd.activationkey
 from xmlrpc import client as xmlrpclib
-
-
-@pytest.fixture
-def shell():
-    """
-    Create fake shell.
-    """
-    base = MagicMock()
-    base.session = hashlib.sha256(str(time.time()).encode("utf-8")).hexdigest()
-    base.client = MagicMock()
-    base.client.activationkey = MagicMock()
-    base.do_activationkey_list = MagicMock(return_value="do_activation_list")
-
-    return base
+from helpers import shell
 
 
 class TestSCActivationKey:
